fix: EncryptedDataKey providerInfo
The providerInfo is technically bytes.
That most keyrings chose to use store this data as a string just convenience.
It is easy and manageable to store pass utf8 strings around,
however some keyrings may chose to use this field to store binary data.
The raw AES keyrings are a notable example.
To complicate matters, utf8 is "destructive" because of multi-byte characters.
binary != decodeUtf8(encodeUtf8(binary))
Any binary value above 127 will be interpreted as a multi-byte character.
To support the simplicity of strings but the extensibility of binary
I chose default to strings, but offer an optional binary property.
All serialize/deserialize operations will prefer the binary value if present.
*It is not required that the providerInfo string "equal" the binary rawInfo*