feat: maxBodySize constraint on decrypt
#48

On decrypt the Encryption SDK **must not** return unverified plaintext.
This means that each “frame” must be buffered before being returned.
In the case of non-framed messages the entire message must be buffered.
This implies that a memory DoS is possible if someone submits a large message.
This constraint protects against this case.