fix: Default CMM should not be required
#44, #70

Since the Default CMM is a defined specification,
it should never be necessary to create an instance of it.
The name should include Default in the name.
Keyrings should always Just Work when being passed
All examples and usages should use Keyrings.

To support the above a bit more work than expected was required.
Primarily in MultiKeyring.
Getting instances to play nicely required code refactoring.
Nothing was added,
but more importantly,
*no existing tests for MultKeyrings were changed*.

The rest of this is changing the names
and making sure Keyrings work.