fix: encrypt/decrypt node
Small frames exacerbate a race between VerifyStream and DecipherStream.
Once an `AuthTag` event is sent to DecipherStream
the current frame is decrypted and verified before being flushed.
More frame data will not be processed by DecipherStream
because DecipherStream will not execute the _transform callback
until the frame is flushed.
However nothing stops VerifyStream from continuing to process data.
VerifyStream will push the next frame's data,
which will be cached by DecipherStream.
But once VerifyStream reaches the end of the frame it will emit `AuthTag`.
This means the DecipherStream will receive the AuthTag before the frame data
and it will correctly fail on precondition checks.

To resolve this a simple callback is passed from VerifyStream to DecipherStream
in the `AuthTag` event.

Tests for both encrypt and decrypt are included here,
because that is how this issue was found.