fix: aws sdk version dependencies pollution
resolves #136
resolved #138

The AWS SDK for JS v2 and v3 have similar types,
but they are not exactly the same.

In a effort to have both versions work,
I imported as dev dependencies all the relevant packages
and tried to use conditional types.
However, since the types are all exported by the functions
this did not work as planed and caused _every_ client to want _every_ SDK.

The interface the AWS Encryption SDK for JS needs
from the AWS SDK is minimal.
Hand rolling the appropriate types is the best solution at this time.