feat: Node.js Typescript version dependency
resolves #135

The AWS Encryption SDK should track the Node.js LTS policy.
For developers using Typescript and @types/node@8
the ESDK should “just work”.
Because some packages export Node.js types like `CipherGCM`
developers would be forced to install an unwanted type version.

This pulls in the needed parts of the needed types into the packages.
It is only the types that differ in this case.

Regarding KeyObjects, rsa support is removed,
but will be added back in when #74 is addressed.