fix: sequence number order
The sequence number **must** be honored.
This is especially important for non-signed algorithm suites.

The test vector added to this PR illustrates this.
It is the encrypted string `’asdf’` with a frame length of 1.
The frames have been rearranged to `’fdsa’`.
Without this change the ESDK would decrypt this blob.