fix: browser encrypt signature encoding
resolves: #154

DER encoding is always signed.
Browsers use 'raw' encoding.
Where the r and s values are just concatenated.
If the first bit of r or s is 1,
an additional 0 byte should be added.
The current implementation of raw2der does not do this.

Tests and test vectors.
The basic solution was to let the asn1 library handle the encoding
and not pass bytes.