fix: ENCODED_SIGNER_KEY in encryption context
A CryptographicMaterialsManager can change entries to the encryptionContext
but changing these values has consequences.
The DefaultCryptographicMaterialsManager uses the value in the encryption context
to store public signing key.
If the caller is using this value in their encryption context the Default CMM
is probably not the CMM they want to use.