fix: integration silent errors
The cli’s for the integration packages could silently error.
The logic is wrapped in an immediately invoked function expression.
The function is async, and if there is an error
this results in an unhandled promise rejection.
Catch the error, log, and exit with a non-zero code.

I also cleaned up some of the path expressions,
and removed the ts-ignore.