fix: Update user agent calls to KMS
We want to be able to track the user agent used in calls made with this SDK.
This is to be able to identify which version of the Encryption SDK is being used
to assist in any future debugging requested by customers.

This tracking is done in the default `clientProvider`.
To opt out of this, you can create your own.