fix: updating readElement to match underlying data structure
The readElement function was previously not taking advantage of
the underlying data structure, which is a sequence of elements,
each of which is a sequence of fields. This resulted in several
superfluous calls to built-in functions (e.g. Array.prototype.splice).

This commit adds an additional parameter to readElement, which is
the number of fields per element, and adapts the code accordingly.
The return of readElement is now an array of arrays, which simplifies
its manipulation by the two caller functions.