chore: Prep for dependabot
`lerna link convert`

This links all the packages to the root
and removed the need for `lerna bootstrap`
This moves all the `devDependencies` from the individual
`package.json` files to the root,
and links the internal modules together.
This does make things complicated
for `integration-browser` and `integration-node`
because these files compile the TS to JS for a CLI.
But this JS file does not exists because it is build from the Typescript.
So there is a chicken/egg situation where the file needs to exists to be installed,
but the file needs to be built.
I added a hack to `touch` the file when the monorepo is built.

This should help with dependabot PRs
as well as making `npm audit` just work.