chore: update dependency versions
This is an explicit update to versions.
Additionally all sub modules package-lock.json files are removed.

When `lerna link` was run, in #242
all modules were linked together at the root.
This means that the root package-lock.json file is responsible
for all sub modules state.
Lerna does not maintain the state on these modules package-lock.json anymore.
So this information is redundant.

Finally, update `local_verdaccio_publish`
to not reset package-lock.json files as they do not exist.