fix!: The AWS Kms keyring use of clientProvider
On encrypt if an AWS KMS client can not be obtained
Encrypt should HALT.

There MUST be a client for every KeyId
When a user configures a AWS KMS keyring with IDs
the intent is to be able to independently decrypt
with ANY of these IDs.
See: https://github.com/awslabs/aws-encryption-sdk-specification/blob/master/framework/kms-keyring.md#onencrypt-goal
Without a client it is
impossible to encrypt under the given CMK.

BREAKING CHANGE:

The previous behavior was to
allow the client supplier to modify
the AWS KMS KeyId intent.
This resulted in a successful
encrypt but WITHOUT the filtered KeyIds