feat: Add maxHeaderSize
Resolves: #388

Add maxHeaderSize feature to control
the maximum size of a header that is allowed.

Processing ESDK messages
of unknown size introduces a risk
of highly-bounded resource consumption
due to variable length fields
in the message format.
This is bad for users
that need to decrypt such messages,
especially for such users
that care about
the memory usage of their systems.

For more details see:
https://github.com/awslabs/aws-encryption-sdk-specification/blob/master/proposals/2020-06-26_decrypt-max-header-size-max-body-size/proposal.md